+++ .gitignore
@@ -1,10 +1,14 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
+# Ignore all, necessary for ignoring binaries
+*
 
-# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
-# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
-Cargo.lock
+# Unignore all with extensions
+!*.*
 
-# These are backup files generated by rustfmt
-**/*.rs.bk
+# Unignore all dirs
+!*/
+
+#ignore local logs
+logs/
+
+#multistage not implemented
+/*.multistage
