+++ README.md
@@ -1,1 +1,9 @@
 # RS_Dinning_Hall
+
+Laboratory written by Boicu Stefan FAF 203
+Prerequisites for building the services: Docker installed and running on the machine. https://docs.docker.com/get-started/
+
+To build the dinning hall image Run 
+'$docker build -t hall-image .' if you are on windows
+or  
+'$sudo docker build -t hall-image .' if you believe in Linux supremacy
